+++ tests/integration/test_receive_ca_cert_status.py
@@ -4,7 +4,6 @@
 
 """Regression test for receive-ca-cert status after relation removal (#670)."""
 
-import os
 import time
 from pathlib import Path
